Allow type annotations
So far we have concluded the GraphQL types strictly from a corresponding
Java type. But in some cases, the underlying data type is the same, only
the usage differs.

This introduces a "Currency" type which is expressed a long value
interpreted as 1/10000th currency units.

We rename the input annotation to @GraphQLField to be usable for both
input and output types.

We add a type() value that can be used to annotate an ambiguous java
type like long into the "Currency" type.

We test that for POJO based output types and input types.

We still need a way to configure this "from the outside" for generated
JOOQ domains.