Fix Input type conversion
The input conversion for domainql methods was mostly working, but turns
out I never actually tested complex object input. While the scalars
involved where converted allright, the complex objects remained as maps.

This introduces a BiMap pojoClass => InputType and vice versa to the
generation process.

With this we add the missing conversion to GraphQLValueProvider.

Add tests for both analysis and execution.