Implement output getter args
So far we've supported data-fetching via svenson properties, i.e.
following the Java bean property semantics.

GraphQL further allows to add arguments to all fields and we support
that by allowing getters with @GraphQLField annotation to have
parameters that we map to GraphQL arguments like we do for queries and
mutations.

This adds an implementation for getter arguments and adds two test cases
for it.