Optionally support @full directive
There still can be situations in which the limitations of a GraphQL schema are
too limiting and the overhead of JSONifying a complete response object
isn't that bad anyway.

This is for example the case with svenson type look ahead based
jsonification that could only be supported with lots of union
definitions and listing all possible fields in the query.

This commit provides an escape-hatch out of the GraphQL type system that
does not change the rest of the development work-flow too drastically.

We implement an optional directive @full that can be used when querying
@GraphQLQuery or @GraphQLMutation methods. Instead of following through with
GraphQL fetching logic, the return value of those methods is returned
as-is inside a DomainQLExecution instance that must be provided as
.context() in the GraphQL endpoint.

Service Configuration
=====================

The DomainQL service definition must be configured to support the @full
directive

    final GraphQLSchema schema = DomainQL.newDomainQL(dslContext)

        ...

        .withFullDirectiveSupported(true)
        .buildGraphQLSchema();

Query/Mutation Setup
====================

The query or mutation declaration must give a full=true parameter on the
annotation.

    @GraphQLQuery( full = true)
    public FullResponse fullQuery()
    {
        ...
    }

The GraphQL queries/mutations themselves must use the @full directive

    query fullQuery
    {
        fullQuery @full
    }

The schema will declare the @full queries/mutations as returning a
boolean which will always be true.

Endpoint Setup
==============

The end-point must provide an instance of
de.quinscape.domainql.DomainQLExecutionContext as .context() of the
ExecutionInput.

    ExecutionInput executionInput = ExecutionInput.newExecutionInput()
        // language=GraphQL
        .context(domainQLExecutionContext)
        .build();

The actual full result can be retrieved by calling
domainQLExecutionContext.getResponse() on that instance.

This commit also provides some test-cases for the new directive.