Refactor type handling and add degenerification
While developing the degeneration feature I ran into some issues with
the existing code that had bothered me before but now proved me right
in being a real obstacle to the degeneration feature.

The code grew out of an exploratory prototype without a preexisting plan
besides creating a workflow model for graphql-java and a mix of
annotation-based and convention-based paradigma interfacing well with
the kind of POJOs generated by JOOQ.

This refactoring does a lot of restructuring that should have been there
earlier in that it separates operations into two phases.

First the detection of the involved java types and following java
properties and now generic types to fill out the final graph of Java
types being translated into the GraphQL schema. This commit introduces
intermediary types InputType and OutputType to gather that knowledge
about java types and the context in which they appear.

The TypeContext class expresses that type context as it is drawn out of
the generic runtime type information of the coressponding getter and
parameter types. For a class Container<C> there can be multiple
instances of that class which all will be degenerified into their
own container type substituting the generic types with types resolved
from the type context.

Domain centric vs Query/Mutation centric
========================================

Another thing that has occured to be before is that there is a more
elegant approach to get to that set of java types with corresponding
GraphQL types we need.

When I started, I naturally thought of the GraphQL domain as being
identical to the JOOQ Domain, which seemed just natural.

On the other hand, it is very natural to have only parts of a domain
being exposed to users with a certain set of roles. Or for application
use cases only being a slice of a larger data context.

Now it may be vital to generate all the JOOQ classes we need for the
application to handle the backend functionality, but that doesn't mean
that all of those JOOQ generated classes are actually exposed to the
client.

The better way to go at this is to simply use the configured set of
JOOQ tables and to follow the input and ouput types used in GraphQL
queries and mutations.

This makes the whole "mirror input" and "override input" functionality
void. If you use your JOOQ domain class "Foo" as input type, there will
be a "FooInput". If you use another "Foo", that one will be used as
"FooInput".

This commit doesn't create input mirrors anymore, it just uses the java
input types referenced and creates GraphQL input types for them. Each
input type will end in "Input". If the java class already ends with
"Input", no further "Input" will be added.

Degenerification
================

So far we have supported Java generics in so far as we followed svenson
in supporting collections of complex types and, in constrast to the
older svenson lib, could actually do so without additionally annotation
just by reflecting on for example the concrete `List<Pojo>` types.

This commit adds a more complex way of dealing with generics that is
degenerification. The initial, very common use-case goes like this.

We have a generic container class Paged to contain a current page of
objects and an row count of the available objects so the UI knows
how many pages there are etc.

```java
public class Paged<T>
{
    private List<T> rows;

    private int rowCount;

    public Paged()
    {
        this(null, -1);
    }

    public Paged(List<T> rows, int rowCount)
    {
        this.rows = rows;
        this.rowCount = rowCount;
    }

    @NotNull
    public List<T> getRows()
    {
        return rows;
    }

    public void setRows(List<T> rows)
    {
        this.rows = rows;
    }

    @NotNull
    public int getRowCount()
    {
        return rowCount;
    }

    public void setRowCount(int rowCount)
    {
        this.rowCount = rowCount;
    }
}

```

Before this commit we were just stuck here and couldn't really express
this in GraphQL without enumerating all possible types and then having
even worse queries that repeat every field of every possible type
(*yuck*).

This commit now allows us to degenerify `Paged` automatically for every
return of it in a query or mutation method.

```java

    @GraphQLQuery
    public Paged<Payload> getPayload()
    {
        final Paged<Payload> paged = new Paged<>();

        // ...

        return paged;
    }

```

This will result in a virtual type "PagedPayload" being created.

```graphql
type PagedPayload {
  # List.rows
  rows: [Payload]!

  # int.rowCount
  rowCount: Int!
}
```

By default the new name for the virtual type is created by concatenating
the generic name and the type variables.

The name can be configured by annotating the query or mutation method with
the new @ResolvedGenericType annotation

```java

    @GraphQLQuery
    @ResolvedGenericType("PagedAndRenamed")
    public Paged<AnnotatedPayload> getPayload()
    {
        // ...

        return paged;
    }

```

This commit adds two query type tests and one execution test for the
degenerification feature.