Add generic domain support, addit. scalars/inputs
While the GraphQL schema world is wonderful when working with concrete
types (even more wonderful with degenerification), it is a bit of an
obstable when trying to implement generic operations that can act on
any kind of domain type. There are no input unions, even if we liked
that option.

What we do is to define a custom scalar for that that transports the
map data as-is to the client into a generic domain object and then
adds an extra conversion step in GraphQLValueProvider to convert the
domain object properties to the right Java type. This double conversion
strategy is unfortunately needed due to the inability of a scalar to
access the schema it is part of.

This scalar is based on a DomainObject interface that adds a "_type"
property referencing the original type name and some generic methods to
read properties from and write properties to these types generically.
(getProperty(name), setProperty(name,value) and propertyNames())

Additional scalar types
=======================

To pull this whole feature off, we need two more features. First we need
to fix that so far, we've only been working with a fixed set of scalar
types without the ability to define new ones.

So this adds the ability to define new scalar types based on a Java type
and a GraphQLScalarType implementation. There can be multiple scalar
types defined for a single Java type in which case the actual type
has to be provided with a @GraphQLField annotation.

Additional Input Types
======================

Another feature I had already considered and found potentially useful.

We need the ability to define input types beyond the types accepted by
our end points.

The reason for that here is, that we might just get along with the new
generic functionality to store a domain type "Foo" without ever
accepting that Foo as query or mutation parameter. (The generic end
points just use "DomainObject" parameters)

Another reason this might be useful is to define organizing container
types for the client-side without ever actually accepting them as input.

Requirements
============

For the Domain Object scalar to work, it has to be added to the GraphQL
schema

    import de.quinscape.domainql.generic.DomainObject;
    import de.quinscape.domainql.generic.GraphQLDomainObjectScalar;

    // ...

    final GraphQLSchema schema = DomainQL.newDomainQL(null)
        .objectTypes(Public.PUBLIC)
        .withAdditionalScalar(DomainObject.class, new DomainObjectScalar())

        // ...

        .buildGraphQLSchema();

In addition to that, all the domain object types we want to use
generically have to implement the DomainObject interface.

For the POJOs generated JOOQ this can be achieved by configuring the new

    de.quinscape.domainql.generic.DomainObjectGeneratorStrategy

as generator strategy in the pom.xml.

Usage
=====

In the end, we can define a mutation method like

    @GraphQLMutation
    public boolean store(DomainObject domainObject)
    {
        // ...
    }

The object must have the "_type" property signalling the target domain
type name. The conversion will result in the paramter `domainObject` to
be passed as the generated POJO for the Foo type.