Introduce DomainQLAware interface
I still wasn't happy with the way the DomainObjectScalar was forced to
do a two-phase conversion due to not being able to access the type
registry and the GraphQLSchema in question.

The easiest solution is just two create a marker interface for a
two-phase intialization. First the scalar is created with a delegating
coercion object which then can be registered with the final DomainQL
and GraphQLSchema instances before we release conrol of the latter.

This makes every registered scalar implementing DomainQLAware to be
called with those final values.