Support jsonb
Adds several classes to deal with Postgresql jsonb fields in a DomainQL
schema.

Adds a JSONB class which is a container for JSON object data (*no*
top-level arrays) that is a svenson DynamicProperties object and stores
data in an internal map. It can be constructed from a map, there's an
.asMap() method to access a read-only version of the internal map.
Adds a forValue(String) method to parse a JSONB instance from a json
string. Adds a toJSON() method to convert it back to json.

Adds a JSONScalar that converts the field into/from a Map.

Adds a PGJSONBinding class to use as field type for JOOQ generated
POJOs for json and jsonb fields.