Implement type-parameterized queries/mutations
While we're already supporting generified types for container types, we
can use another way to use generics within the GraphQL world.

This commits adds the ability to express a special kind of generic
GraphQL query/mutation

Let's take a look at an example from the tests:

    @GraphQLQuery
    public <T> T query(
        @GraphQLTypeParam(types = { TypeA.class, TypeB.class }) Class<T> cls,
        ComplexInput complexInput
    ) throws IllegalAccessException, InstantiationException
    {
        // ...
    }

The interesting part is the @GraphQLTypeParam annotated first parameter.

This example will produce two different queries, one for each of the
listed types TypeA and TypeB.

By default the names will be "queryTypeA" and "queryTypeB", but you can
substitute another name pattern.

At runtime, the expected return type is passed in the `cls` parameter.

This commit adds tests for queries with return values `T`,
`Container<T>` and `List<T>`, which is what we want to support here.

Refactoring
===========

To accomplish this, there was a need to refactor type-context away from
the very method centric way it was still implemented so that we can
substitute the same parent context for one of the @GraphQLTypeParam
provided types.

The new code seems much cleaner in any case.