Fix weird issues with parametrized mutations
The type params commit did not actually test type parametrized mutations.
While the code is pretty much the same, it never hurts to actually test
it, so I started adding some type parametrized mutation tests and got
the weirdest errors. Tried updating graphql-java, which seemed to change
the errors to a similarly confusing but at least reproducible errors.

    Validation error of type VariableTypeMismatch:
    Variable type 'ComplexInput' doesn't match expected type 'ComplexInput'
    @ 'queryTypeB', locations=[SourceLocation{line=3, column=30}],
    description=

    'Variable type 'ComplexInput' doesn't match expected type 'ComplexInput'

Yikes.

After some debugging I came upon the fact that this message occured
because the second ComplexInput was in fact an unresolved GraphQLTypeReference

So what we're doing now is preventing this from happening by rearranging
the order in which we are creating the final GraphQL types and
remembering the created complex types so that when we create the queries
and mutations last, we can resolve the type references to the actual
types before we hand them to graphql-java.

This commit also contains some random minor cleanups I did while searching for
the error.