Fix TypeContext edge case
The way we created the TypeContext for methods that return the parameter
type of a parameterized query directly was wrong.

    public <T> T query(...)
    {
        ...
    }

The method has Object.class as non-generic type with a type map
mapping T to the actual type.

This bug was covered up by the fact that when we ran the tests, the
discovery order of query/mutation methods is not reproducible and
varies. And if the first parameterized method discovering our
TypeA/TypeB test types happened to bei returning either Container<T> or
List<T>, the resulting type context would be correct and correctly
analyze TypeA/TypeB.

When the Object.class error then occured, our caching of already created
types would prevent that erroneous type context to take full effect.

Only when the direct returning method was discovered first and thus
analyzed first, we had fixed around the wrongly generated "ObjectTypeA"
type before and thus we were generating a GraphQL type "TypeA" but
analyzed Object.class as its type, finding no properties to map to
GraphQL leading to errors due to the missing "value" fields.

Hard bug to understand.

This commit changes the behavior to detect our special case and
assigning the right type as base type and generating the right name for
it.

The former member field `degenerifyName` is not really needed and
refactored into a parameter to buildTypeName.