Refactor Test: Assert empty errors
When I started writing the execution test I checked for the errors very
awkwardly by asserting the length of the errors to be 0. While this is
an okay thing to do for normal code, doing so here creates the
problem that I want to know what errors there are if I assume there are
none in the test. Which is why I then added lots of log.info statements
to log them etc.

This commits instead changes the assert to assume the errors to be equal
an empty list. If that assertion fails, hamcrest/IntelliJ will present
us with the error output automatically.