Refactor: Property handling
Protect more iterations over the svenson properties from readOnly or
ignored properties.

This gets rid of the _type fields for our DomainObject which we don't
need anymore at this point.

This commit changes the assert to account for the changing number of
properties from this change.

This commit changes some accesses to fields by index to a named access
which seems less brittle and also better to read.