Make DomainQL accessible from env.
For internal data fetchers and queries/mutations it can be very helpful
to access the current DomainQL instance without making the caller
provide it again. Since we're already wrapping the
DataFetchingEnvironment, it is easy to add the current DomainQL instance
to that wrapper.

Adds a test that injects our DomainQLDataFetchingEnvironment directly
and tests for the presence of a DomainQL instance.