Provide information about generic types
For the degenerified types it can be very useful to reliably know at
runtime which GraphQLType is based on which generic type and also maybe
what the type parameters are.

This commit tracks the creation of degenerified output types and
collects a GenericTypeReference for each.

The type reference contains the name of the original generic type, the
name of the generated degenerified GraphQL type and the type arguments
as GraphQL types.

This commit changes the SchemaDataProvider to add the collected
references as "genericTypes" to the "__schema" root returned by the
introspection query. This behavior is enabled by default and can be
disabled by calling the three-argument constructor.

Changes the schema data provider test to check for generic types.