Implement GenericScalar
Like we did for the DomainObject scalar, there are situations where we
want to tunnel a generically typed scalar value while still using the
normal scalar conversions for it.

This commit implements a GenericScalarType that adds a "GenericScalar"
scalar type and must be registered with

    .withAdditionalScalar(
        GenericScalar.class,
        GenericScalarType.newGenericScalar()
    )

The java side type for this scalar is the GenericScalar bean.

```java
package de.quinscape.domainql.generic;

import java.util.Objects;

public class GenericScalar
{
    private String type;

    private Object value;

    public GenericScalar()
    {
        this(null, null);
    }

    public GenericScalar(String type, Object value)
    {
        this.type = type;
        this.value = value;
    }

    public String getType()
    {
        return type;
    }

    public void setType(String type)
    {
        this.type = type;
    }

    public Object getValue()
    {
        return value;
    }

    public void setValue(Object value)
    {
        this.value = value;
    }

    // ...

}
```

The type property contains the name of the GraphQL scalar type and value
a valid java side value for this scalar.

The JSON representation for GenericScalar is a map object with type and
value properties. The value property will contain the serialized JSON
representation of that scalar type.

    {
        "type" : "Int",
        "value" : 1234
    }

Adds a test case for GenericScalar handling