Introduce FetcherContext
We offer a simple yet effective way of optimizing GraphQL query
fetching. The idea is to allow a fetcher to store data for another
fetcher that would otherwise be needed to fetch in a different query or
as part of a batched query.

We add a JSON-ignored property fetcherContext to our DomainObject
interface and the two implementations.

The reference fetcher will look if the domain object it is fetching its
id from has such a fetcherContext property.

If it has a fetcherContext, the reference fetchers returns a property of
that context, that has the same name as the original GraphQL field for
which the reference fetcher is registered.

Meta-Data
---------

To support the preparation of these fetcher contexts, we add some more
meta data to the data fetchers.

Since we can't really attach our own domainql related data to the
GraphQL schema, the data fetchers are a convenient location to store
some additional information about the field fetched that, apart from
being used for the function of the fetcher can also serve as information
source for secondary analysis, e.g. when constructing optimized queries.

We rename SvensonFetcher to FieldFetcher because the name was just bad.

We might have to add more meta-data later. The current information is
based on the needs of the implementation in automaton.