Add @GraphQLScalar annotation
Forgetting to register a scalar implementation can easily happen,
especially with transititive dependencies.

Trying to analyze such a scalar implementation as GraphQL pojo leads to
strange errors.

So we add an annotation @GraphQLScalar that marks such scalar
implementations.

If we encounter such an @GraphQLScalar annotated class in the
TypeRegistry as POJO type, it was obviously not registered as scalar, so
we throw an exception notifying the user.