Add safety check for scalar names
Accidental name collisions lead to very confusing bugs, especially if
they happen to randomly use one or the other type.

We can mitigate this by doing a safety check when we create our
scalarByName map.

We check if there is a scalar type already defined for each name and if
it's not the identical type (The java types "int" and "Integer" both use
the identical GraphQL "Int" type)

If there is such a conflict, we throw a speaking exception.