Refactor Domain Object property handling
We already have the "_type" property to deal with and the fetcher
context introduced another property we don't really want.

Failing to handle the "_type" property for the most parts has little
effect, it's just another string property set / transmitted.

JSONifying the fetcher context e.g. however does not work at all.

We make sure DomainObject.propertyNames() method does ignore "_type" and
we evade all the problems with FetcherContext by renaming the setter and
getter methods to no longer obey the property rules. The main point is
attaching data to a domain object and retrieving it again in a later
DataFetcher.

This commit renames the methods to `provideFetcherContext` and
`lookupFetcherContext`.
git