Add support for database views
So far, the database backed parts of the DomainQL schema relied totally
on what the JOOQ generator generated for an SQL schema. JOOQ however
does not seem to support generating pojos based on view shapes.

This commit changes the DomainQL constructor to take the table lookup
map we used to create later in the constructor and move the creation of
the lookup into the builder.

This allows adding a new method

    DomainQLBuilder.objectType(String viewName, Class<?> cls)

where viewName is the name of an SQL table-like that can be queried like
a table and cls is the POJO type to use for that SQL name.

The given POJO must match the result shape exactly and must have each
column property annotated with @Column annotations and can be annotated
with @NotNull, just like JOOQ generator generates it.

an @Table annotation must be given to define the name of the table-like.

    import javax.persistence.Table;
    import javax.persistence.Column;
    import javax.validation.constraints.NotNull;

    @Table(name = "sum_per_month")
    public class SumPerMonth
    {
        private int month;
        private int year;
        private int sum;

        @Column(name = "month")
        @NotNull
        public int getMonth()
        {
            return month;
        }

        // ...
    }

Annotations on the class itself are just ignored at this point.

With this commit, Domain users can lookup the JOOQ tables and fields of
these view POJOs and use them for queries.