Implement schema doc extraction and injection
So far we've autogenerated the description parts of the DomainQL schema
which was mostly useless as it only reflected Java types and DB Tables
(that latter maybe with some merit)

We remove the very redundant description of Java-originated fields that
really only reflected the GraphQL type in Java and was adding only
noise.

We add a system of optional documentation extraction and injection.

de.quinscape.domainql.docs.DocsExtractor that extracts GraphQL centric
documentation from our @GraphQLLogic classes and Java-side enums / POJO
types corresponding to GrapihQL types.

It a hierarchy starting with de.quinscape.domainql.docs.TypeDoc, here as
GraphQL types (which they normally are not)

    #Generated for de.quinscape.domainql.docs.TypeDoc
    type TypeDoc {
      description: String
      fieldDocs: [FieldDoc]
      name: String
    }

    #Generated for de.quinscape.domainql.docs.FieldDoc
    type FieldDoc {
      description: String
      name: String
      paramDocs: [ParamDoc]
    }

    #Generated for de.quinscape.domainql.docs.ParamDoc
    type ParamDoc {
      description: String
      name: String
    }

This data can be written out to files or whatever and later be imported
into the generated DomainQL domain with three new methods in
DomainQLBuilder

    .withTypeDocs(Set<TypeDoc>)
    .withTypeDocsFrom(File)
    .withTypeDocsFrom(InputStream)

the latter two are convenience methods to parse a JSON file or input
stream containing JSON data into an Array of TypeDoc instances.

The schema generation will look up descriptions based on type names and
field names and argument names.

If it finds a description in the imported type docs, it uses that
description as is, otherwise it uses the revised default method of
generating descriptions.

Type References in degenerified Types
=====================================

Degenerified types can be documented by either writing and providing
specific documentation for the degenerified name.

If no such type documentation exists, the documentation of the generic
type they're based on is used.

The type name in square brackets is replaced with the name of the
parameter type. For example a description "List of [T] rows" is replaced
with "List of Foo rows" in `PagedFoo`.