Refactor DocsExtractor
We need stable, reproducible output to minimize changes in git and to
make the changes only happen when something changes etc pp.

Using Set<TypDoc> feels more and more like a charade. Obviously the
ordering of types is important, so we skip the "incidentally
insertion-ordered" and replace the Set<TypeDoc> with List<TypeDoc>.

Operation complexity as sucky as before, slight improvement on object
creations maybe with ArrayList.

We sort the fields in QueryType and MutationType and we sort the types
in special ordering with QueryType and MutationType first.