Change reference renaming
In some cases it is desirable to rename the generated fields for foreign
key references.

The implementation so far was broken in that the OBJECT_AND_SCALAR mode
with a defined left side name would try to generate both the scalar
field and the object field under the new name, which resulted in the
scalar field to be overwritten.

The scalar names are both unique and far more likely to be a correct
representation of the value (as opposed to using a value name to refer
to the target object).

So we change the behavior to always rename the object and to ignore the
naming parameters for any mode not involving object field creation.