JSONHolder: Allow constructing from json
For now we've been using JSONHolder to jsonify objects early but we can
use the same principle to directly construct JSON output from Java
objects that never really existed. For example using the JSONBuilder to
create a partial JSON snapshot of a a java side map without creating the
intermediary map on the java side.

The use cases are so closely related that we can safely make them one.

This adds a JSONHolder(String) constructor that takes the given JSON
string as-is.