Check for field name conflicts
With an unfortunately designed database schema and the default rules (
 e.g. having both "thing" and "thing_id" as fields) by spefically naming
fields wrong it used to be possible for reference and back reference
objects to be defined with the same name as an already existing database
field of the same name.

This kind of works as GraphQL just overwrites the old field definition
in the newer versions, but is certainly not something we want to have or
support.

So this commit adds safety checks in this regard.

Adds two tests to test a conflict for reference and back reference
objects and fixes some cases in which the test code already produced
such a name conflict.