Refactor relation handling
So far we used two methods .configureRelation() with different
signatures to create a RelationConfiguration containing some additional
configuration in addition to the database foreign key auto-generated by
JOOQ.

This only allowed us to use relations for true database foreign keys and
nescessitated figuring out further data based on the generated foreign
keys in multiple code locations. We were also not able to use relations
with database views.

While it is true that database views allow you to join together
arbritrary information and run aggregate functions on it etc pp, all
this information adds up as flattened properties within a new
view-specific type, losing the normal GraphQL abilities for the
underlying base data.

By being able to define a relation based on a database view POJO type,
we can just include the object id fields in the view and work with the
normal GraphQL types from there on. Instead of necessitating the
definition of many views per use, we can use a single database view in
multiple ways.

Another weakness of the current code was the structure of the relation
configuration. The more complicated variant of that method already took
5 parameters and further extending those seemed definitely like a bad
idea.

This commit deprecates the two methods and adds a new
.withRelation(RelationBuilder) that accepts a sub-builder for a
refactored relation model that contains the JOOQ tables and fields as
well as the POJO type information about classes and property names.
While we're at it, we add support for multi-key foreign keys.

Changes all the old tests to use the new config and adds new test cases
for multi-key foreign keys and POJO-based view relations.

Adds the Postgresql database dump used to create the generated test
database schema classes and adds 2 new tables and a new view for the
tests.

Increases the project version to 0.2-SNAPSHOT due to the deprecation,
possible breaking changes.

Reference object names
----------------------

Determines reference object names early.

So far we kept the default values of leftSideObjectName and
rightSideObjectName in the relation model and only resolved the actual
name when creating the field.

It is better to do the resolving early so that the relation model always
contains the actual names of the underlying reference and back reference
fields.