Provide unique relation model ids
Make sure that every relation model has a unique id, whether the user
defines one or not. As default use `SourceType-objectFieldName` where
`SourceType` is the name of the leftside domain type and
`objectFieldName` the name of the left side object field or the first
foreign key name if there is no left side object.

User-provided ids must be unique, default ids are made unique by
appending a counter.

Adds a test case for the relation model ids.