Fix input list conversion
We actually had no tests for having List types as inputs for our end
points and the projects only used List<DomainObject> so far, which is a
list of *scalar values* which is not quite the same.

Trying to have a list of concrete domain objects as input would complain
about not being able to cast a LinkedHashMap to the concrete type
because the conversion never happened.

During the implementation I also stumbled upon enums being potentially
another source of misbehavior and actually found another tiny bug.

Fixes conversion of list input types and adds three new test cases for
List<DomainObject>, List<ComplexInput> and List<MyEnum> as input types.