Implement name fields
This commit introduces the concept of name fields to DomainQL.

Name fields are fields that are representative for an domain entity.
Most commonly, they're a field called "name", but the field can have
other names and there might be more than one of them.

In fact, some entities might only make sense within in a context.

Configuration
-------------

Adds two new method in DomainQLBuilder which allow the definition of
those name fields per type.

    b.configureNameFields( Bar.class,"name", ...)

configurNameFields defines multiple fields as representative of a type.
It allows field definition in a dot notation.

A GraphQL type like

    Bar
    {
        name
        owner{
            name
        }
    }

can user "owner.name" as one of the fields.

As a convenience method, we also added another method that allows the
definition of a name field for multiple types.

    .configureNameFieldForTypes("name", BarOwner.class, Foo.class)

This method requires less configuration and we expect it to be the more
common use case.

The domainql object will validate all type and field definitions on
creation.

Access
------

The domainQL instance then offers access to that configuration via

    domainQL.getNameFields()

which returns a map of domain type names mapping to the list of name
fields for that type.