Fix GenericScalar serialization
Serialization of GenericScalar values was broken in that it used the
wrong conversion function, which does not matter for a lot of the
scalars where JSON-format and Java format are identical.

Adds a test for the complex use case we discovered this with which was
serializing a GenericScalar with a list of domain object scalars as
value.