Fix null-handling in DomainObject scalar
When using DomainObject defined as scalar value we had a wrong handling
of null-values.

Nullable fields could not contain null values as the serialization was
broken. Most likely related to the previous serialization problem in
that this never happened before.

So we make sure that we leave null values in nullable fields as-is and
complain about finding them in non-nullable ones.

Adds two test-cases for that.

Suppresses the errors that are logged here.