Add auto name field configuration
So far we had the methods to explicitly define all name fields for all
types and multiple name fields for a type.

This adds a third, more convenient way to configure name fields and that
is by name.

    .configureNameField("name")

now allows to define one or multiple name fields. Whenever an output
type has such a name field, we register that field as name field for the
type.

Explicit type configuration still has predence over that.

This complicates DomainQL object creation a bit, but I think it's worth
it in terms of convenience.