Provide read-only access to the JOOQ table lookup
It can be helpful to know if a domain type has originated from a
database table or not. getJooqTable() allows access to the table but
errors if we feed it a non-DB type name.

This method now allows enumeration over all DB types as well as
convenient access to their table lookups.