Handle simple array properties
While it is certainly not a really meaningful use-case, we might just
encounter types containing binary data and we should at least be able to
accept those into the GraphQL world.

This prevents the type analyzer from going astray in the register phase
with arrays and then we accept fields with arrays of all scalar types.

Most commonly this will be byte[] which then ends up as GraphQL type
[Byte].

Adds two tests to dealing with a bean containing a byte[] property.