Refactor BigDecimal/BigInteger handling
The current state of BigDecimal/BigInteger handling does not make much
sense to me. Sure, on a pure JSON level, any syntactically correct
number can be represented correctly, but most commonly, as it is in our
case, we need to parse the JSON and don't want to write a custom JSON
parser just to deal with the parsing of big numbers beyond the scale or
exactness provides by JavaScript Numbers.

And if I would to implement such a parser, what would I store the result
as? I would have to define a standard way to use BigDecimal/BigIntegers
in the app and deal with the many different precisions that can be
present in a database.

The much more simple approach seems to be to just convert BigDecimal and
BigInteger values to strings within the wire format.

The values can be safely transported and might even be stored in an
application without turning them into actual numbers when we imagine an
data input application with all calculations taking part in the backend.

But we also can deal with the numeric values on a case by case basis and
convert them into the bignumber.js or whatever values with the
appropriate scale and precision.

Removes the registration of the default graphql-java scalar
implementations for BigDecimal and BigInteger. Applications are free to
disagree with me and manually register them again, that's fine.

Adds two scalar implementations based on a string wire format
which can be registered alternatively.

Adds a schema test cases and an execution test case.