Upgrade graphql-java
We upgrade our mother library to 16.2 and have to deal with some
changes.

This removes the deprecated .dataFetcher() and .argument() calls
and changes the dataFetcher handling to the new GraphQLCodeRegistry
functionality.

GraphQLType has no name field anymore so there's a lot of casting to
GraphQLNamedType now.

Types produced with notNull() now seem to *not* be equal to other types
with the same spec, something we used extensively in the tests so far.

This adds converts the types to strings and then changes the asserts to
expect the String notation of the type instead of recreating the type.

We now introduce dummy methods when we have no queries or no mutations,
which, while exotic doesn't seem impossible to have either a write-only
service for a static HTML site or a read-only service for a website.