Implement @GraphqlComputed
Allow the introduction of computed properties that are both available on
the client side as well as on the server.

The functionality rests on being able to provide an alternate
Implementation for the generated database types.

The test case uses an override for the TargetSeven type of our test
database.

import de.quinscape.domainql.annotation.GraphQLComputed;

public class TargetSeven
    extends de.quinscape.domainql.testdomain.tables.pojos.TargetSeven
{

    @GraphQLComputed
    public String getConcat()
    {
        return this.getId() + ":" + this.getName();
    }
}

it's basically just a class that has the same simple name as one of the
generated types and that extends that type to add new properties etc.

Using that type as output in an GraphQL method then enables the override
to take place and the system uses the extended version of that type
instead of the generated version.

There is a bit of refactoring an cleanup done in the order of
registering types and how it works internally. This was initially
thought of as preparatory refactoring, but the desired functionality
emerged naturally from it.

The types of the GraphQL methods are processed first now and win with
their defined overriding types, be it for input types or output types.

For an overriding output type, the system will now first detect it as
return type of a GraphQL method and register the type for that concrete
Java type. When the database mechanism later resolves the type it will
use the extended java type and the inheritance will provide all the
columns as they are. Note that it is not strictly necessary to use
inheritance, it's just the easiest thing to do here.

You can just as well mirror all properties of the type in the same
manner that JOOQ generates them, i.e. with all the annotations DomainQL
reacts to.

The computed property is not considered writable and is not mirrored in
input types.

For the output type, we mark the field definition in the GraphQL schema
with a special @computed directive.